CDK-917: Combine merge-schemas and schema command.
This removes the merge-schemas command and adds a --merge option to the
schema command.
